requirements.txt
supabase>=2.0.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
numpy>=1.24.0
streamlit>=1.30.0
PyMuPDF>=1.23.0
fastapi>=0.100.0
uvicorn>=0.23.0 